Deploy a highly available web application using CloudFormation
Initial commit.
Creating this project will give you the hands-on experience you need to confidently talk about infrastructure as code. So, for that reason, we have chosen a realistic scenario where you deploy an application (Apache Web Server) and you also pick up code (JavaScript and HTML) from S3 Storage and deploy it in the appropriate folder on the web server.

Added:
* Network infrastructure
* Serevers infrastructure